+++ src/bin/ginfo/main.rs
@@ -1,0 +1,12 @@
+mod cli;
+mod commands;
+
+use anyhow::Result;
+use clap::Parser;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    let args = cli::GInfoCli::parse();
+    commands::handle_commands(&args.config, args.command).await?;
+    Ok(())
+}