+++ src/bin/gjob/main.rs
@@ -6,7 +6,7 @@ use clap::Parser;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let args = cli::GCtl::parse();
+    let args = cli::GJob::parse();
     commands::handle_commands(&args.config, args.command).await?;
     Ok(())
 }